refactor(compare): extract C8 to checks/visit_n.py (Phase A 9/N)
Move check_c8_visit_distribution and its _synthesize_source_visit_counts
helper into the checks subpackage. visit_n imports Path and yaml directly
since the synthesizer reads visit.yaml from disk.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1237 → 1051 lines.
